rename runs_config to config
